feat: 修复SVPWM算法 + 新增正负序分解模块

+++ Components/Transforms/svpwm.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <stdint.h>
+#include "fast_math.h"
 
 typedef struct {
     float vbus;
@@ -13,10 +14,10 @@ typedef struct {
     float v_beta;
     float v_d;
     float v_q;
-    float sector;
-    float ta;
-    float tb;
-    float tc;
+    int sector;           /**< 扇区号 (0-5)，使用整数避免浮点精度问题 */
+    float ta;             /**< A 相占空比 */
+    float tb;             /**< B 相占空比 */
+    float tc;             /**< C 相占空比 */
 } SvPwm_Handle;
 
 void SvPwm_Init(SvPwm_Handle *h, const SvPwm_Config *cfg);
